Allow interface start without reload.
If an interface runs a bridge, it may be undesirable to have the
interface reloaded at runtime. At the same time, it may be desirable to
start the interface if it's not yet started.

Add start/start_type attributes mirroring reload/reload_type, so that
you can define new interfaces with reload=false, start=true, and have
then started but not reloaded.

As a related cleanup, add a config_file attribute to debian interfaces,
to refactor the repeated path.

Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>